requirements.txt
agent-framework==1.0.0b251120
agent-framework-core==1.0.0b251001
aiofiles==25.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.11.0
attrs==25.4.0
azure-core==1.36.0
azure-core-tracing-opentelemetry==1.0.0b12
azure-identity==1.25.1
azure-monitor-opentelemetry==1.8.2
azure-monitor-opentelemetry-exporter==1.0.0b45
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
-e file:///home/luis-cm/Documents/Github/chatbot-civic
click==8.3.1
cryptography==46.0.3
distro==1.9.0
ecdsa==0.19.1
fastapi==0.121.3
googleapis-common-protos==1.72.0
grpcio==1.76.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
importlib-metadata==8.7.0
isodate==0.7.2
jiter==0.12.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
mcp==1.22.0
msal==1.34.0
msal-extensions==1.3.1
msrest==0.7.1
oauthlib==3.3.1
openai==2.8.1
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-instrumentation==0.59b0
opentelemetry-instrumentation-asgi==0.59b0
opentelemetry-instrumentation-dbapi==0.59b0
opentelemetry-instrumentation-django==0.59b0
opentelemetry-instrumentation-fastapi==0.59b0
opentelemetry-instrumentation-flask==0.59b0
opentelemetry-instrumentation-psycopg2==0.59b0
opentelemetry-instrumentation-requests==0.59b0
opentelemetry-instrumentation-urllib==0.59b0
opentelemetry-instrumentation-urllib3==0.59b0
opentelemetry-instrumentation-wsgi==0.59b0
opentelemetry-proto==1.38.0
opentelemetry-resource-detector-azure==0.1.5
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
opentelemetry-semantic-conventions-ai==0.4.13
opentelemetry-util-http==0.59b0
packaging==25.0
passlib==1.7.4
protobuf==6.33.1
psutil==7.1.3
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.4
pydantic-core==2.41.5
pydantic-settings==2.12.0
pyjwt==2.10.1
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.29.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sse-starlette==3.0.3
starlette==0.50.0
tenacity==9.1.2
tqdm==4.67.1
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.5.0
uvicorn==0.38.0
websockets==15.0.1
wrapt==1.17.3
zipp==3.23.0
